requirements.txt
streamlit
numpy
pandas
scikit-image
scikit-learn
opencv-python
altair
joblib
Pillow